requirements.txt
fastapi
uvicorn
python-multipart
uvicorn[standard]
google-cloud-aiplatform
google-generativeai
pillow